separated nodes and agent into separate files after improvements to their features; Rewrote the README to fully describe the persona
